Further updates to seeding pipeline
